Compress demo bundles and split provenance lens
